refactor and update dependencies

+++ scripts/deploy.sh
@@ -1,11 +1,19 @@
 #!/usr/bin/sh
 
-unlink site-active
+# Remove current active site link
+[ -L site-active ] && unlink site-active
+
+# Pull latest image and extract generated static files
 docker pull keeb/keeb.dev
-docker run --name tmp-keeb-dev-deploy keeb/keeb.dev echo true 
-# || docker rm tmp-keeb-dev-deploy
-docker cp tmp-keeb-dev-deploy:/hexo/keeb.dev/public .
+docker run --name tmp-keeb-dev-deploy keeb/keeb.dev echo "deployment ready"
+docker cp tmp-keeb-dev-deploy:/app/public .
 docker rm tmp-keeb-dev-deploy
+
+# Create timestamped deployment
 mv public $(date +%F)
 ln -s $(date +%F) site-active
-docker-compose restart
+
+# Restart services if using docker-compose
+if [ -f docker-compose.yml ]; then
+    docker-compose restart
+fi
